Extraction of networking logic on QuesitonsList screen
